requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
bs4==0.0.2
python-dotenv==1.0.1
telebot==0.0.5
pyTelegramBotAPI==4.25.0
Telethon==1.38.1
pillow==11.1.0